refactor: mv computation basis as separated module
